added narrow version of dashboard layout
