+++ src/react-ui/layouts/index.tsx
@@ -1,1 +1,3 @@
 ﻿export { default as Dashboard } from './Dashboard.tsx';
+export { default as DashboardNarrow } from './DashboardNarrow.tsx';
+export { default as DashboardShared } from './DashboardShared.tsx';
